requirements.txt
torch
torchvision
numpy
scipy
pandas
matplotlib
seaborn
scikit-learn
umap-learn
hdbscan
imbalanced-learn
xgboost
catboost
medmnist
ipykernel
